chore(infrastructure): "Add support for Poetry dependency management"
Makes the safe-autonomy-dynamics repo a python-library ACT 3 project that utilizes Poetry for dependency management. Some python-library template documentation files are added, but not filled out or incorporated into the existing documentation. The existing documentation was updated to discuss setup and installation via Poetry rather than via pip.
