+++ .gitlab/issue_templates/Bug.md
@@ -1,0 +1,30 @@
+## Summary
+
+<!-- Summarize the bug encountered concisely -->
+
+## Steps to reproduce
+
+<!-- How one can reproduce the issue - this is very important -->
+
+## Example Project
+
+<!-- If possible, please create an example project here on GitLab.com that exhibits the problematic behavior, and link to it here in the bug report. If you are using an older version of GitLab, this will also determine whether the bug has been fixed in a more recent version -->
+
+## What is the current bug behavior?
+
+<!-- What actually happens -->
+
+## What is the expected correct behavior?
+
+<!-- What you should see instead -->
+
+## Relevant logs and/or screenshots
+
+<!-- Paste any relevant logs - please use [code blocks](https://about.gitlab.com/handbook/markdown-guide/#fenced) to format console output, logs, and code. -->
+
+## Possible fixes
+
+<!-- If you can, link to the line of code that might be responsible for the problem -->
+
+/label ~Bug
+/cc @primaryauthor
