+++ .gitlab/issue_templates/Task.md
@@ -1,0 +1,19 @@
+## Summary
+
+<!-- Summarize the desired feature/behavior concisely -->
+
+## Completion Criteria
+
+<!-- In this section, list the criteria for the task to be considered complete - i.e. This task can be considered complete when tool-x can perform action a, b, and c -->
+
+- [ ] Implementation
+- [ ] Unit testing
+- [ ] Documentation
+
+## Relevant link/resources
+
+<!-- Paste any relevant logs - please use [code blocks](https://about.gitlab.com/handbook/markdown-guide/#fenced) to format console output, logs, and code. -->
+
+<!-- If there are resources that you know of that would provide a good starting point for the person implementing this task, consider adding them here -->
+
+/cc @primaryauthor
