+++ .gitlab/issue_templates/User Story.md
@@ -1,0 +1,7 @@
+## Story
+
+As a user/admin/dev/ops, I want to be able to do-this-or-that for this-reason.
+
+/label ~"User Story"
+
+/cc @primaryauthor
