+++ .gitlab/issue_templates/default.md
@@ -1,0 +1,28 @@
+## Issue Summary
+
+<!--Summarize the reason for creating this issue. Typically this is a sentence concisely describing the desired feature behavior or the problem being fixed.-->
+
+## Completion Criteria
+
+<!--Complete the sentence below and list the criteria for the task to be considered complete. -->
+
+This issue is defined as done when ...
+
+<!--Example: This task can be considered complete when tool-x can perform action a, b, and c-->
+
+### Developer Checklist
+
+<!--Mark items in the checklist below as completed before closing this issue. Delete any items that do not apply to this issue.-->
+
+- [ ] Code implementation completed
+- [ ] Unit testing completed
+- [ ] Code comments are written/updated
+- [ ] New issue is created for tech writer to create/update documentation
+
+<!-- The logic for prompting a developer to create a separate issue for work assigned to the technical writer is to expedite the ability to close the original issue when the developer completes their work. This aligns to the team philosophy about closing issues as quickly as possible.-->
+
+## Relevant Links/Resources
+
+<!-- Paste any relevant logs - please use [code blocks](https://about.gitlab.com/handbook/markdown-guide/#fenced) to format console output, logs, and code. -->
+
+<!-- If there are resources that you know of that would provide a good starting point for the person implementing this task, consider adding them here -->
