+++ .markdownlint.yaml
@@ -1,0 +1,6 @@
+# Default state for all rules
+default: true
+
+MD013:
+  # Number of characters
+  line_length: -1
