+++ .markdownlintignore
@@ -1,0 +1,3 @@
+
+/CHANGELOG.md
+/.gitlab/issue_templates/*
