+++ docs/developer-guide.md
@@ -1,0 +1,19 @@
+# safe-autonomy-dynamics Developer Guide
+
+## Design Patterns
+
+<!-- Describe code organization and style -->
+
+## Testing
+
+### Unit Tests
+
+<!-- Describe how to run unit test -->
+
+### Functional Tests
+
+<!-- Describe how to run functional tests -->
+
+## Releasing
+
+The act3-pt CLI contains a `act3-pt release` command that automates this process.
