+++ docs/quick-start-guide.md
@@ -1,0 +1,36 @@
+# safe-autonomy-dynamics Quick Start Guide
+
+## Purpose
+
+<!-- In this section, we tell the user what to expect in the quick start documentation for this product/process.
+
+This should be a 1-2 sentence summary of the documentation. The first sentence is provided below. Add additional content if needed. -->
+
+This Quick Start Guide is for users who have already installed safe-autonomy-dynamics and who are ready to take advantage of its features..
+
+You will be guided through the steps necessary to begin using safe-autonomy-dynamics
+<!-- Include a Table of Contents that links to steps below if possible -->
+
+1. [Configuration](#configuration)
+2. [Initial Usage](#initial-usage)
+
+## Configuration
+
+<!-- This section will begin with a summary of the overall initial configuration process or set up steps a user must complete to begin interacting with the product.-->
+
+<!-- ### First sub-section
+
+This sub-section should be clearly labeled and describe the first step in the configuration or set up process that must be completed.
+
+### Additional sub-section(s)
+
+Add as many sub-sections as needed.-->
+
+## Initial Usage
+
+<!-- Provide enough information in this section that a new "cold" user could begin using the product. Refrain from providing a deep dive into features, concepts, and other in-depth content that will be in the User Guide.-->
+
+## Additional Resources
+
+- [Documentation](../README.md#documentation)
+- [Support](../README.md#support)
