+++ docs/troubleshooting-faq.md
@@ -1,0 +1,27 @@
+# safe-autonomy-dynamics Troubleshooting FAQ
+
+<!--This document is a place to capture questions that have come up repeatedly by existing users and that can be answered in a helpful manner through written documentation. The contents included here should be generalizable enough to apply to all users with the same or similar questions. This is not an appropriate place to document issues or questions that are individualized to a particular environment or use case.-->
+
+## <!--Section 1--> Troubleshooting
+
+<!--Use section headings to group similar types of questions together. For instance, you might rename Section 1 to Initial Configuration Troubleshooting -->
+
+Question 1
+<!--Replace the text Question 1 with the language of the first FAQ-->
+> Answer 1
+<!--Replace the text Answer 1 with the corresponding answer/solution. Do not remove the formatting to display this text as a pull quote. Repeat the question and answer pattern as many times as necessary-->
+
+<!-- markdownlint-disable-next-line -->
+## <!--Section 2--> Troubleshooting
+
+<!--Don't forget to update the section heading. For instance, you might rename Section 2 to Initial Usage Troubleshooting -->
+
+Question 1
+<!--Replace the text Question 1 with the language of the first FAQ-->
+> Answer 1
+<!--Replace the text Answer 1 with the corresponding answer/solution. Do not remove the formatting to display this text as a pull quote. Repeat the question and answer pattern as many times as necessary-->
+
+## Additional Resources
+
+- [Documentation](../README.md#documentation)
+- [Support](../README.md#support)
