+++ docs/user-guide.md
@@ -1,0 +1,40 @@
+# safe-autonomy-dynamics User Guide
+
+## Purpose
+
+<!--This section should begin with a purpose statement that clearly defines the intended user and their purpose for using the ACT3 project or product. The purpose statement should be as clear and succinct as possible.-->
+
+safe-autonomy-dynamics is the primary way for \<person or process> to \<complete action>.
+
+<!--Example of a well written Purpose statement:
+Project Tool is the primary way for ACT3 developers to generate new projects and update existing projects by adopting ACT3's best practices. Project Tool is also the primary way for ACT3 developers to begin using the ACT3 pipeline.-->
+
+<!-- If applicable, include this sentence: All (related products or processes) use (product/process/feature), so to (summary of related actions) it is crucial/important/necessary to understand (this product/process/feature name). -->
+
+On this page:
+
+<!-- Provide a Table of Contents to help users jump directly to a desired section -->
+- [Overview](#overview)
+- [Concepts](#concepts)
+- [Usage](#usage)
+
+## Overview
+
+<!-- This section should provide a high level overview that helps the user construct a mental model of what the project/product is and how it works. The goal of this section is to prepare the reader for understanding the concepts defined and described in the section below. Diagrams and flow charts can be integrated here when appropriate.  -->
+
+## Concepts
+
+<!-- The Concepts section helps readers learn about the parts of the ACT3 project/product and the abstractions ACT3 uses to represent key concepts. 
+
+This section of the User Guide should help readers obtain a deeper understanding of how this particular project/product leverages key concepts to accomplish the user's end-goal(s).
+
+Related concepts from other projects or products may need to be included in this section to help the user develop a mental model of this project/product.-->
+
+## Usage
+
+<!-- This section should be used to build on the key concepts described above and help the user understand how to practically apply the concepts to accomplish their objective(s). Refrain, if possible, from getting into specific commands or detailed discussion of functionality. Those items should be documented separately and linked from the Documentation section of the README file  -->
+
+## Additional Resources
+
+- [Documentation](../README.md#documentation)
+- [Support](../README.md#support)
