+++ pyproject.toml
@@ -21,11 +21,11 @@ packages = [
     ]
 
 [tool.poetry.dependencies]
-python = ">=3.9, <3.13"
+python = ">=3.9,<3.13"
 numpy = "^1.23.5"
 jax = { version = "0.4.3", extras = ["cpu"], optional = true }
-scipy = ">=1.10.0"
-pydantic = "^1.9.2"
+scipy = "^1.11.2"
+pydantic = "^2.3.0"
 pint = "^0.22"
 
 [tool.poetry.extras]
