+++ renovate.json
@@ -1,0 +1,6 @@
+{
+    "extends": [
+        ":semanticCommits",
+        ":semanticCommitType(fix)"
+    ]
+}
