--- requirements-compliance.txt
@@ -1,25 +1,0 @@
-# linters
-mypy==0.930
-mypy-extensions==0.4.3
-mypy-protobuf==2.7
-types-PyYAML==5.4.11
-flake8==3.9.2
-pylint==2.8.3
-
-#testers
-pytest==6.2.2
-pytest-cov==2.11.1
-pytest-order==0.10.0
-flake8==3.9.2
-
-# formatters
-yapf==0.31.0
-isort==5.9.3
-rope==0.19.0
-
-# pre-commit
-pre-commit==2.13.0
-pre-commit-hooks==4.0.1
-detect-secrets==1.1.0
-blacken-docs==1.12.1
-bashate==2.0.0
