--- requirements-jax.txt
@@ -1,1 +1,0 @@
-jax[cpu]==0.4.3
