--- requirements-mkdocs.txt
@@ -1,17 +1,0 @@
-# docs requirements
-mkdocs
-mkdocs-git-revision-date-localized-plugin
-mkdocs-material
-mkdocs-material-extensions
-mkdocs-mermaid-plugin
-mkdocs-pdf-export-plugin
-mkdocstrings==0.18.0
-mkdocs-autorefs
-mkdocs-coverage
-mkdocs-gen-files
-mkdocs-section-index
-pytkdocs[numpy-style]>=0.5.0
-
-papermill
-Pillow
-pymdown-extensions
