+++ setup.cfg
@@ -11,10 +11,9 @@
 [mypy]
 # plugins = pydantic.mypy
 ignore_missing_imports = True
+show_error_codes = True
 files=safe_autonomy_dynamics
 
-# [mypy-ray.rllib.*]
-# follow_imports = skip
 
 ###################
 # flake8 sections #
@@ -26,7 +25,7 @@ files=safe_autonomy_dynamics
 # W504: line break after binary operator
 # C901: Function is too complex
 ignore = E266, W503, W504, C901
-exclude = tests/*
+exclude = test/*
 max-line-length = 140
 max-complexity = 18
 select = B,C,E,F,W,T4
@@ -79,7 +78,7 @@ max-line-length=140
 # For a list of configurations go to https://docs.pytest.org/en/stable/customize.html
 
 [tool:pytest]
-testpaths = tests
+testpaths = test
 addopts = --strict-markers -m "not slow" --disable-pytest-warnings
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
