+++ test/test_simulators/test_entities/scripts/cwh_rotation_dynamics_for_test_cases.py
