+++ test/test_simulators/test_entities/scripts/cwh_rotation_x_thrust_case_to_y_thrust_case.py
